Update version 2

+++ routes/web.php
@@ -1,9 +1,11 @@
 <?php
 
 use App\Cafe;
+use App\Facility;
 use App\Http\Controllers\CafeController;
 use App\Http\Controllers\CategoryFoodController;
 use App\Http\Controllers\EventController;
+use App\Http\Controllers\FacilityController;
 use App\Http\Controllers\GaleryController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\Location;
@@ -37,27 +39,32 @@ Route::get('/detail/{cari}',[NewsController::class, 'show']);
 Route::get('/detail/{news}/edit',[NewsController::class, 'edit'])->middleware('is_admin');
 Route::get('/eventlist',[NewsController::class, 'showData'])->middleware('is_admin');
 Route::put('/event/edit/{news}',[NewsController::class, 'update'])->middleware('is_admin');
-
 Route::get('/event/create',[NewsController::class, 'createindex'])->middleware('is_admin');
 Route::post('/event/create',[NewsController::class, 'post'])->middleware('is_admin');
-
-Route::delete('/event/post/{news}',[NewsController::class, 'destroy']);
+Route::delete('/event/post/{news}',[NewsController::class, 'destroy'])->middleware('is_admin');
 
 Route::get('/galery',[GaleryController::class,'index']);
-
-Route::post('/galery/upload',[GaleryController::class,'store']);
-Route::delete('/galery/delete/{galery}',[GaleryController::class, 'destroy']);
+Route::post('/galery/upload',[GaleryController::class,'store'])->middleware('is_admin');
+Route::delete('/galery/delete/{galery}',[GaleryController::class, 'destroy'])->middleware('is_admin');
 
 Route::get('/location',[Location::class, 'index']);
 
 Route::get('/cafe',[CafeController::class,'index']);
-Route::get('/cafe/create',[CafeController::class,'create']);
-Route::post('/cafe/create',[CafeController::class,'store']);
-Route::delete('/cafe/delete/{cafe}',[CafeController::class,'destroy']);
-Route::get('/cafe/{cafe}/edit',[CafeController::class,'edit']);
-Route::put('/cafe/edit/{cafe}',[CafeController::class,'update']);
-Route::post('/cafe/filter',[CafeController::class,'filter']);
+Route::get('/cafe/create',[CafeController::class,'create'])->middleware('is_admin');
+Route::post('/cafe/create',[CafeController::class,'store'])->middleware('is_admin');
+Route::delete('/cafe/delete/{cafe}',[CafeController::class,'destroy'])->middleware('is_admin');
+Route::get('/cafe/{cafe}/edit',[CafeController::class,'edit'])->middleware('is_admin');
+Route::put('/cafe/edit/{cafe}',[CafeController::class,'update'])->middleware('is_admin');
 
 Route::get('/category',[CategoryFoodController::class,'index']);
-Route::post('/category/create',[CategoryFoodController::class,'store']);
-Route::delete('/category/delete/{categoryFood}',[CategoryFoodController::class,'destroy']);
+Route::post('/category/create',[CategoryFoodController::class,'store'])->middleware('is_admin');
+Route::delete('/category/delete/{categoryFood}',[CategoryFoodController::class,'destroy'])->middleware('is_admin');
+
+Route::get('/facility',[FacilityController::class, 'index']);
+Route::get('/facility/list',[FacilityController::class, 'list'])->middleware('is_admin');
+Route::get('/facility/create',[FacilityController::class, 'create'])->middleware('is_admin');
+Route::post('/facility/create',[FacilityController::class, 'store'])->middleware('is_admin');
+Route::get('/facilty/detail/{cari}',[FacilityController::class, 'detail']);
+Route::get('/facilty/{facility}/edit',[FacilityController::class, 'edit'])->middleware('is_admin');
+Route::put('/facility/edit/{facility}',[FacilityController::class, 'update'])->middleware('is_admin');
+Route::delete('/facility/delete/{facility}',[FacilityController::class, 'destroy']);
